refactor: split olf.release into manifest/compatibility/readiness modules
Decompose the 873-line flat olf/release.py into a package capability-
split into _manifest (component manifest/checksums), _compatibility
(compatibility-matrix rendering and drift checks), and _readiness (the
olf release check gate), per issue #101. Tests reorganized to mirror
the new boundaries; no CLI behavior change.
